scripts/requirements-llm.txt
# Python dependencies for LLM position extraction
# Install in python-embeddings container

psycopg2-binary>=2.9.9
requests>=2.31.0
